Add -w option to sosetup for generating an answer file
This patch enables a -w <file> option to sosetup that generates a new
answerfile that can then be fed into sosetup with the -f <file> option.
It specifically does not support sosetup-network, and warns the user
that the network must be configured before running it with the answer
file it generates.  It does support running in generate mode as a non-
root user.